API Change: changed the code location to int64 instead of Ints. Moved to a ResourceConsumer architecture that can consume both streamed source code, or paths for file processors that want to handle their own file parsing

+++ pkg/code/model.go
@@ -4,14 +4,14 @@ package code
 type Position struct {
 
 	//Line position in a document (zero-based).
-	Line int `json:"line"`
+	Line int64 `json:"line"`
 
 	//Character offset on a line in a document (zero-based). Assuming that the line is
 	//represented as a string, the `character` value represents the gap between the
 	//`character` and `character + 1`.
 	//If the character value is greater than the line length it defaults back to the
 	//line length.
-	Character int `json:"character"`
+	Character int64 `json:"character"`
 }
 
 //LessOrEqual returns whether `thisPosition` is chronologically "before" `thatPosition`
